refactor: requestMapping 클래스 추가
