+++ app/src/main/java/nextstep/jwp/JwpApplication.java
@@ -1,12 +1,13 @@
 package nextstep.jwp;
 
-import nextstep.jwp.http.WebServer;
+import nextstep.jwp.infrastructure.RequestMapping;
+import nextstep.jwp.infrastructure.WebServer;
 
 public class JwpApplication {
 
     public static void main(String[] args) {
         int port = WebServer.defaultPortIfNull(args);
-        final WebServer webServer = new WebServer(port);
+        final WebServer webServer = new WebServer(port, new RequestMapping());
         webServer.run();
     }
 }
