+++ app/src/main/java/nextstep/jwp/infrastructure/WebServer.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.infrastructure;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,9 +15,11 @@ public class WebServer {
     private static final int DEFAULT_PORT = 8080;
 
     private final int port;
+    private final RequestMapping requestMapping;
 
-    public WebServer(int port) {
+    public WebServer(int port, RequestMapping requestMapping) {
         this.port = checkPort(port);
+        this.requestMapping = requestMapping;
     }
 
     public void run() {
@@ -34,7 +36,7 @@ public class WebServer {
     private void handle(ServerSocket serverSocket) throws IOException {
         Socket connection;
         while ((connection = serverSocket.accept()) != null) {
-            new Thread(new RequestHandler(connection)).start();
+            new Thread(new RequestHandler(connection, requestMapping)).start();
         }
     }
